requirements.txt
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
langgraph
akshare
pandas
numpy
matplotlib
python-dotenv
fastapi
uvicorn
ta
faiss-cpu
sentence-transformers
pypdf
aiohttp
requests
PyJWT
bcrypt